map drawing abstraction
removed a lot of hardcoded drawing. map can now be moved around the screen to a degree, with the mouse input staying consistent with visuals. "support" for possible moving around of game window while running due to lack of compulsory actor method call on stage addition. also may be possible to resize entire map to an extent.